betty format

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -11,12 +11,11 @@ int sum_them_all(const unsigned int n, ...)
 {
 	va_list sum;
 	unsigned int i;
-	double total = 0.0;
+	int total = 0;
 
 	va_start(sum, n);
-        if (n == 0)
+	if (n == 0)
 		return (0);
-
 	for (i = 0; i < n; i++)
 		total = total + va_arg(sum, int);
